py
my py is working and now i think i need to move all this to my final. does recieve messages form mqtt and indicates with py -m Server.main that it is trying to send over however can cause it is not operational. so needs tech 5 (final assignment ) that is what is needed fo rnow